django_demo/requirements.txt
Django==2.1
djangorestframework==3.9.1
